feat: Add status, mic position, and reset support
Status just makes calls to get_<feature> functions and prints them out
Position is based on HID calls and hardcoded values I found in pcaps
Reset is very similar to unplugging and replugging the device

Additionally I added a detection for whether an interface was attached
to a kernel driver, and a drop implementation will reattach all such
interfaces to the kernel. My understanding is that `drop()` will be
called even during panic unwinding, and therefore "guarantees" that we
will properly free these interfaces and reattach them to the kernel.